deeplink/v4-webview/requirements.txt
# tested with Python 3.11.4
pywebview==4.2.2
py2app==0.28.6
qtpy==2.3.1
pyqt6==6.5.1
PyQt6-WebEngine==6.5.0